src/main/java/org/zui/model/Game.java
package org.zui.model;

public class Game {
    private Player player1;
    private Player player2;
    private GameStatus gameStatus;

    public Game(Player player1, Player player2) {
        this.player1 = player1;
        this.player2 = player2;
        this.gameStatus = GameStatus.READY_TO_START;
    }

    public Player getPlayer1() {
        return player1;
    }

    public Player getPlayer2() {
        return player2;
    }

    public GameStatus getGameStatus() {
        return gameStatus;
    }

    public void setGameStatus(GameStatus gameStatus) {
        this.gameStatus = gameStatus;
    }
}
